+++ ext/fs/std_fs.rs
@@ -119,10 +119,10 @@ impl FileSystem for RealFs {
     spawn_blocking(move || chmod(&path, mode)).await?
   }
 
-  fn fchmod_sync(&self, fd: i32, mode: u16) -> FsResult<()> {
+  fn fchmod_sync(&self, fd: i32, mode: u32) -> FsResult<()> {
     fchmod(fd, mode)
   }
-  async fn fchmod_async(&self, fd: i32, mode: u16) -> FsResult<()> {
+  async fn fchmod_async(&self, fd: i32, mode: u32) -> FsResult<()> {
     spawn_blocking(move || fchmod(fd, mode)).await?
   }
 
@@ -351,7 +351,7 @@ fn chmod(path: &Path, _mode: u32) -> FsResult<()> {
 }
 
 #[cfg(unix)]
-fn fchmod(fd: i32, mode: u16) -> FsResult<()> {
+fn fchmod(fd: i32, mode: u32) -> FsResult<()> {
   use nix::sys::stat::fchmod;
   use nix::sys::stat::Mode;
   let mode: Mode = Mode::from_bits_truncate(mode);
@@ -361,7 +361,7 @@ fn fchmod(fd: i32, mode: u16) -> FsResult<()> {
 
 // TODO: implement fchmod for Windows
 #[cfg(not(unix))]
-fn fchmod(fd: i32, mode: u16) -> FsResult<()> {
+fn fchmod(fd: i32, mode: u32) -> FsResult<()> {
   Err(FsError::NotSupported)
 }
